+++ Everyone_Is_Willy/Everyone_Is_Willy.csproj
@@ -48,7 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="All_Willy.cs" />
+    <Compile Include="Everyone_is_Willy.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
